Revert "int uint comparison fixes in shaderparam.cpp"
This reverts commit b0c7447a72c6876f5f6557b3a12adc36173c9119.

+++ src/engine/render/shaderparam.cpp
@@ -122,13 +122,13 @@ LocalShaderParamState *LocalShaderParam::resolve()
     {
         return nullptr;
     }
-    if(!(static_cast<int>(s->localparamremap.size()) > loc))
+    if(!(s->localparamremap.size() > loc))
     {
         if(loc == -1)
         {
             loc = getlocalparam(name);
         }
-        if(!(static_cast<int>(s->localparamremap.size()) > loc))
+        if(!(s->localparamremap.size() > loc))
         {
             return nullptr;
         }
